requirements.txt
opencv-python
argparse
pillow
numpy
cryptography